LPS-134903 Set editor mode from the properties sidebar
