+++ modules/apps/journal/journal-web/src/main/resources/META-INF/resources/ddm_template_editor/components/App.js
@@ -14,9 +14,8 @@
 
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 
-import {useChannel} from '../hooks/useChannel';
 import {AppContextProvider} from './AppContext';
 import {ClosableAlert} from './ClosableAlert';
 import {Editor} from './Editor';
@@ -32,37 +31,12 @@ export default function App({
 	showLanguageChangeWarning,
 	templateVariableGroups,
 }) {
-	const [editorMode, setEditorMode] = useState(initialEditorMode);
-	const inputChannel = useChannel();
-
-	useEffect(() => {
-		const modeSelect = document.getElementById(
-			`${portletNamespace}language`
-		);
-
-		const onModeChange = (event) => {
-			setEditorMode(
-				{
-					ftl: 'ftl',
-					vm: 'velocity',
-				}[event.target.value]
-			);
-		};
-
-		if (modeSelect) {
-			modeSelect.addEventListener('change', onModeChange);
-
-			return () => {
-				modeSelect.removeEventListener('change', onModeChange);
-			};
-		}
-	}, [portletNamespace]);
-
 	const [selectedSidebarPanelId, setSelectedSidebarPanelId] = useState(null);
 
 	return (
 		<AppContextProvider
-			inputChannel={inputChannel}
+			editorMode={initialEditorMode}
+			portletNamespace={portletNamespace}
 			propertiesViewURL={propertiesViewURL}
 			templateVariableGroups={templateVariableGroups}
 		>
@@ -90,10 +64,7 @@ export default function App({
 
 					<Editor
 						autocompleteData={editorAutocompleteData}
-						editorMode={editorMode}
 						initialScript={initialScript}
-						inputChannel={inputChannel}
-						portletNamespace={portletNamespace}
 					/>
 				</div>
