+++ modules/apps/journal/journal-web/src/main/resources/META-INF/resources/ddm_template_editor/components/PropertiesSidebarPanel.js
@@ -25,7 +25,34 @@ export default function PropertiesSidebarPanel({className}) {
 	const [content, setContent] = useState('');
 	const isMounted = useIsMounted();
 
-	const {propertiesViewURL} = useContext(AppContext);
+	const {portletNamespace, propertiesViewURL, setEditorMode} = useContext(
+		AppContext
+	);
+
+	useEffect(() => {
+		if (!loading) {
+			const modeSelect = document.getElementById(
+				`${portletNamespace}language`
+			);
+
+			const onModeChange = (event) => {
+				setEditorMode(
+					{
+						ftl: 'ftl',
+						vm: 'velocity',
+					}[event.target.value]
+				);
+			};
+
+			if (modeSelect) {
+				modeSelect.addEventListener('change', onModeChange);
+
+				return () => {
+					modeSelect.removeEventListener('change', onModeChange);
+				};
+			}
+		}
+	}, [loading, setEditorMode, portletNamespace]);
 
 	useEffect(() => {
 		fetch(propertiesViewURL)
